Fix MSVC + CUDA 13.3 build and bump dependencies
Get the project building with MSVC (VS 18 / cl 14.51) and CUDA 13.3:

- Patch spdlog's bundled fmt for NVCC: its device front-end (EDG) never
  sees the host /utf-8 flag, so fmt's compile-time UTF-8 probe misfires,
  and EDG rejects the char32_t rounding-threshold table (hex escapes with
  the high bit set). Guard both behind __CUDACC__ via CPM PATCHES
  (cmake/patches/fmt-nvcc-compat.patch).
- Forward /utf-8 and /Zc:preprocessor to the CUDA host compiler; CCCL
  (thrust/cub) requires the conforming preprocessor. CXX gets /utf-8
  transitively from spdlog's PUBLIC export.
- Fix a Win64 std::max type mismatch in cuda/broad_phase/utils.cu
  (unsigned long is 32-bit on Windows): use size_t(1).

Dependency bumps: spdlog 1.15.3 -> 1.17.0, Eigen -> 5.0.1,
libigl -> 2.6.0, oneTBB 2022.1.0 -> 2022.3.0.

Also add CMakePresets.json, ignore *.vs, and stop leaking /FS to NVCC.

+++ src/scalable_ccd/cuda/broad_phase/utils.cu
@@ -22,7 +22,7 @@ void setup(
 
     if (!boxes_per_thread) {
         boxes_per_thread =
-            std::max(max_shared_memory / sizeof(AABB) / max_threads, 1ul);
+            std::max(max_shared_memory / sizeof(AABB) / max_threads, size_t(1));
     }
     logger().trace("Boxes per thread: {:d}", boxes_per_thread);
